+++ app/client/keybase/debug/keystore.go
@@ -122,9 +122,9 @@ func fetchValidatorPrivateKeysFromK8S() (map[string]cryptoPocket.KeyPair, error)
 }
 
 func fetchValidatorPrivateKeysFromFile() (map[string]cryptoPocket.KeyPair, error) {
+	// BUG: When running the CLI using the build binary (i.e. `p1`), it searched for the private-keys.yaml file in `github.com/pokt-network/pocket/build/localnet/manifests/private-keys.yaml`
 	// Get private keys from manifest file
 	_, current, _, _ := r.Caller(0)
-
 	//nolint:gocritic // Use path to find private-keys yaml file from being called in any location in the repo
 	yamlFile := filepath.Join(current, privateKeysYamlFile)
 	if exists, err := fileExists(yamlFile); !exists || err != nil {