+++ utility/transaction_test.go
@@ -1,4 +1,4 @@
-package test
+package utility
 
 import (
 	"encoding/hex"
@@ -7,11 +7,10 @@ import (
 
 	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	"github.com/pokt-network/pocket/shared/codec"
+	"github.com/pokt-network/pocket/shared/converters"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/crypto"
-	"github.com/pokt-network/pocket/utility"
 	typesUtil "github.com/pokt-network/pocket/utility/types"
-	utilTypes "github.com/pokt-network/pocket/utility/types"
 	"github.com/stretchr/testify/require"
 )
 
@@ -20,47 +19,43 @@ var (
 )
 
 func TestUtilityContext_AnteHandleMessage(t *testing.T) {
-	ctx := NewTestingUtilityContext(t, 0)
+	ctx := newTestingUtilityContext(t, 0)
 
 	tx, startingBalance, _, signer := newTestingTransaction(t, ctx)
-	_, signerString, err := ctx.AnteHandleMessage(tx)
+	_, signerString, err := ctx.anteHandleMessage(tx)
 	require.NoError(t, err)
 	require.Equal(t, signer.Address().String(), signerString)
-	feeBig, err := ctx.GetMessageSendFee()
+	feeBig, err := ctx.getMessageSendFee()
 	require.NoError(t, err)
 
 	expectedAfterBalance := big.NewInt(0).Sub(startingBalance, feeBig)
-	amount, err := ctx.GetAccountAmount(signer.Address())
+	amount, err := ctx.getAccountAmount(signer.Address())
 	require.NoError(t, err)
 	require.Equal(t, expectedAfterBalance, amount, "unexpected after balance")
-
-	test_artifacts.CleanupTest(ctx)
 }
 
 func TestUtilityContext_ApplyTransaction(t *testing.T) {
-	ctx := NewTestingUtilityContext(t, 0)
+	ctx := newTestingUtilityContext(t, 0)
 
 	tx, startingBalance, amount, signer := newTestingTransaction(t, ctx)
-	txResult, err := ctx.ApplyTransaction(0, tx)
+	txResult, err := ctx.applyTx(0, tx)
 	require.NoError(t, err)
 	require.Equal(t, int32(0), txResult.GetResultCode())
 	require.Equal(t, "", txResult.GetError())
-	feeBig, err := ctx.GetMessageSendFee()
+	feeBig, err := ctx.getMessageSendFee()
 	require.NoError(t, err)
 
 	expectedAmountSubtracted := amount.Add(amount, feeBig)
 	expectedAfterBalance := big.NewInt(0).Sub(startingBalance, expectedAmountSubtracted)
-	amount, err = ctx.GetAccountAmount(signer.Address())
+	amount, err = ctx.getAccountAmount(signer.Address())
 	require.NoError(t, err)
 	require.Equal(t, expectedAfterBalance, amount, "unexpected after balance")
-
-	test_artifacts.CleanupTest(ctx)
 }
 
 func TestUtilityContext_CheckTransaction(t *testing.T) {
 	mockBusInTestModules(t)
 
-	ctx := NewTestingUtilityContext(t, 0)
+	ctx := newTestingUtilityContext(t, 0)
 	tx, _, _, _ := newTestingTransaction(t, ctx)
 
 	txBz, err := tx.Bytes()
@@ -71,34 +66,30 @@ func TestUtilityContext_CheckTransaction(t *testing.T) {
 	require.NoError(t, err)
 	require.True(t, testUtilityMod.GetMempool().Contains(hash))
 	require.Equal(t, testUtilityMod.CheckTransaction(txBz).Error(), typesUtil.ErrDuplicateTransaction().Error())
-
-	test_artifacts.CleanupTest(ctx)
 }
 
 func TestUtilityContext_GetSignerCandidates(t *testing.T) {
-	ctx := NewTestingUtilityContext(t, 0)
-	accs := GetAllTestingAccounts(t, ctx)
+	ctx := newTestingUtilityContext(t, 0)
+	accs := getAllTestingAccounts(t, ctx)
 
 	sendAmount := big.NewInt(1000000)
-	sendAmountString := typesUtil.BigIntToString(sendAmount)
+	sendAmountString := converters.BigIntToString(sendAmount)
 	addrBz, er := hex.DecodeString(accs[0].GetAddress())
 	require.NoError(t, er)
 	addrBz2, er := hex.DecodeString(accs[1].GetAddress())
 	require.NoError(t, er)
 	msg := NewTestingSendMessage(t, addrBz, addrBz2, sendAmountString)
-	candidates, err := ctx.GetSignerCandidates(&msg)
+	candidates, err := ctx.getSignerCandidates(&msg)
 	require.NoError(t, err)
 
 	require.Equal(t, 1, len(candidates), "wrong number of candidates")
 	require.Equal(t, accs[0].GetAddress(), hex.EncodeToString(candidates[0]), "unexpected signer candidate")
-
-	test_artifacts.CleanupTest(ctx)
 }
 
 func TestUtilityContext_CreateAndApplyBlock(t *testing.T) {
 	mockBusInTestModules(t)
 
-	ctx := NewTestingUtilityContext(t, 0)
+	ctx := newTestingUtilityContext(t, 0)
 	tx, _, _, _ := newTestingTransaction(t, ctx)
 
 	proposer := getFirstActor(t, ctx, coreTypes.ActorType_ACTOR_TYPE_VAL)
@@ -111,41 +102,37 @@ func TestUtilityContext_CreateAndApplyBlock(t *testing.T) {
 	require.NotEmpty(t, appHash)
 	require.Equal(t, 1, len(txs))
 	require.Equal(t, txs[0], txBz)
-
-	test_artifacts.CleanupTest(ctx)
 }
 
 func TestUtilityContext_HandleMessage(t *testing.T) {
-	ctx := NewTestingUtilityContext(t, 0)
-	accs := GetAllTestingAccounts(t, ctx)
+	ctx := newTestingUtilityContext(t, 0)
+	accs := getAllTestingAccounts(t, ctx)
 
 	sendAmount := big.NewInt(1000000)
-	sendAmountString := typesUtil.BigIntToString(sendAmount)
-	senderBalanceBefore, err := typesUtil.StringToBigInt(accs[0].GetAmount())
+	sendAmountString := converters.BigIntToString(sendAmount)
+	senderBalanceBefore, err := converters.StringToBigInt(accs[0].GetAmount())
 	require.NoError(t, err)
 
-	recipientBalanceBefore, err := typesUtil.StringToBigInt(accs[1].GetAmount())
+	recipientBalanceBefore, err := converters.StringToBigInt(accs[1].GetAmount())
 	require.NoError(t, err)
 	addrBz, er := hex.DecodeString(accs[0].GetAddress())
 	require.NoError(t, er)
 	addrBz2, er := hex.DecodeString(accs[1].GetAddress())
 	require.NoError(t, er)
 	msg := NewTestingSendMessage(t, addrBz, addrBz2, sendAmountString)
-	require.NoError(t, ctx.HandleMessageSend(&msg))
-	accs = GetAllTestingAccounts(t, ctx)
-	senderBalanceAfter, err := typesUtil.StringToBigInt(accs[0].GetAmount())
+	require.NoError(t, ctx.handleMessageSend(&msg))
+	accs = getAllTestingAccounts(t, ctx)
+	senderBalanceAfter, err := converters.StringToBigInt(accs[0].GetAmount())
 	require.NoError(t, err)
 
-	recipientBalanceAfter, err := typesUtil.StringToBigInt(accs[1].GetAmount())
+	recipientBalanceAfter, err := converters.StringToBigInt(accs[1].GetAmount())
 	require.NoError(t, err)
 
 	require.Equal(t, sendAmount, big.NewInt(0).Sub(senderBalanceBefore, senderBalanceAfter), "unexpected sender balance")
 	require.Equal(t, sendAmount, big.NewInt(0).Sub(recipientBalanceAfter, recipientBalanceBefore), "unexpected recipient balance")
-
-	test_artifacts.CleanupTest(ctx)
 }
 
-func newTestingTransaction(t *testing.T, ctx *utility.UtilityContext) (transaction *typesUtil.Transaction, startingBalance, amountSent *big.Int, signer crypto.PrivateKey) {
+func newTestingTransaction(t *testing.T, ctx *utilityContext) (tx *typesUtil.Transaction, startingBalance, amountSent *big.Int, signer crypto.PrivateKey) {
 	amountSent = new(big.Int).Set(defaultSendAmount)
 	startingBalance = new(big.Int).Set(test_artifacts.DefaultAccountAmount)
 
@@ -156,17 +143,17 @@ func newTestingTransaction(t *testing.T, ctx *utility.UtilityContext) (transacti
 	require.NoError(t, err)
 
 	signerAddr := signer.Address()
-	require.NoError(t, ctx.SetAccountAmount(signerAddr, startingBalance))
+	require.NoError(t, ctx.setAccountAmount(signerAddr, startingBalance))
 
-	msg := NewTestingSendMessage(t, signerAddr, recipientAddr.Bytes(), utilTypes.BigIntToString(amountSent))
+	msg := NewTestingSendMessage(t, signerAddr, recipientAddr.Bytes(), converters.BigIntToString(amountSent))
 	any, err := codec.GetCodec().ToAny(&msg)
 	require.NoError(t, err)
 
-	transaction = &typesUtil.Transaction{
+	tx = &typesUtil.Transaction{
 		Msg:   any,
 		Nonce: testNonce,
 	}
-	require.NoError(t, transaction.Sign(signer))
+	require.NoError(t, tx.Sign(signer))
 
 	return
 }