+++ app/client/doc/CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.13] - 2023-02-15
+## [0.0.0.14] - 2023-02-15
 
 - Added `non_interactive` flag to allow for non-interactive `Stake` and `Unstake` transactions (dogfooding in `cluster-manager`)
 - Updated CLI to use to source the address book and the current height from the RPC server leveraging the `rpcAddressBookProvider` and `rpcCurrentHeightProvider` respectively and the `bus` for dependency injection
+
+## [0.0.0.13] - 2023-02-14
+
+- Fixed `docgen` to work from the root of the repository
+- Updated all the CLI docs
 
 ## [0.0.0.12] - 2023-02-14
 
@@ -25,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.10] - 2023-02-07
 
-- Added GITHUB_WIKI tags where it was missing
+- Added GH_WIKI tags where it was missing
 
 ## [0.0.0.9] - 2023-02-06