+++ persistence/docs/CHANGELOG.md
@@ -7,9 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.34] - 2023-02-10
+## [0.0.0.36] - 2023-02-15
 
 - Module now embeds `base_modules.IntegratableModule` for DRYness
+
+## [0.0.0.35] - 2023-02-15
+
+- Add a few `nolint` comments to fix the code on main
+
+## [0.0.0.34] - 2023-02-14
+
+- Remove `IUnstakingActor` and use `UnstakingActor` directly; guideline for removing future unnecessary types (e.g. TxResult)
+- Typo in `GetMinimumBlockHeightQuery`
+- Reduce unnecessary `string` <-> `[]byte` conversion in a few places
+- Fix bug in `updateUnstakedHeightIfPausedBefore` that was unstaking all actors
 
 ## [0.0.0.33] - 2023-02-09