+++ runtime/docs/CHANGELOG.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.17] - 2023-02-10
+## [0.0.0.18] - 2023-02-15
 
 - Added `bootstrap_nodes_csv` in `P2PConfig` to allow for a comma separated list of bootstrap nodes
 - Introduced `modules.ModulesRegistry` for better separation of concerns
 - Added `StateMachineModule` accessors
 - `Manager` embeds `base_modules.IntegratableModule` for DRYness
+
+## [0.0.0.17] - 2023-02-14
+
+- Move shared utils (e.g. `BitIngToString`) to the `converters` package
+- Remove `CleanupTest`
 
 ## [0.0.0.16] - 2023-02-09